Prepare u-boot v2017.09 for rk3588

+++ drivers/demo/demo-simple.c
@@ -1,9 +1,10 @@
-// SPDX-License-Identifier: GPL-2.0+
 /*
  * Copyright (c) 2013 Google, Inc
  *
  * (C) Copyright 2012
  * Pavel Herrmann <morpheus.ibis@gmail.com>
+ *
+ * SPDX-License-Identifier:	GPL-2.0+
  */
 
 #include <common.h>
@@ -16,7 +17,7 @@ static int simple_hello(struct udevice *dev, int ch)
 {
 	const struct dm_demo_pdata *pdata = dev_get_platdata(dev);
 
-	printf("Hello from %08x: %s %d\n", (uint)map_to_sysmem(dev), pdata->colour,
+	printf("Hello from %08x: %s %d\n", map_to_sysmem(dev), pdata->colour,
 	       pdata->sides);
 
 	return 0;
